+++ app_generators/ahn/ahn_generator.rb
@@ -23,7 +23,6 @@ class AhnGenerator < RubiGen::Base
 
       m.file  *["config/environment.rb"]*2
       m.file  *["config/adhearsion.rb"]*2
-      m.file  *["dialplan.rb"]*2
       m.file  *["README"]*2
       m.file  *["Rakefile"]*2
       m.file  *["Gemfile"]*2