+++ app_generators/ahn/templates/dialplan.rb
@@ -1,4 +1,0 @@
-adhearsion {
-  #simon_game
-  speak "hello world"
-}