+++ lib/adhearsion/punchblock.rb
@@ -2,7 +2,6 @@ module Adhearsion
   module Punchblock
     extend ActiveSupport::Autoload
 
-    autoload :Commands
     autoload :MenuDSL
 
     include ::Punchblock