+++ spec/adhearsion/punchblock/menu_dsl/calculated_match_collection_spec.rb
@@ -3,13 +3,18 @@ require 'spec_helper'
 module Adhearsion
   module Punchblock
     module MenuDSL
-
       describe CalculatedMatchCollection do
-        include PunchblockCommandTestHelpers
+        def mock_with_potential_matches(potential_matches)
+          CalculatedMatch.new :potential_matches => potential_matches
+        end
 
-        attr_reader :collection
-        before(:each) do
-          @collection = Adhearsion::Punchblock::MenuDSL::CalculatedMatchCollection.new
+        def mock_with_exact_matches(exact_matches)
+          CalculatedMatch.new :exact_matches => exact_matches
+        end
+
+        def mock_with_potential_and_exact_matches(potential_matches, exact_matches)
+          CalculatedMatch.new :potential_matches => potential_matches,
+                              :exact_matches => exact_matches
         end
 
         it "the <<() method should collect the potential matches into the actual_potential_matches Array" do
@@ -18,11 +23,11 @@ module Adhearsion
           mock_matches_1 = mock_with_potential_matches mock_matches_array_1
           mock_matches_2 = mock_with_potential_matches mock_matches_array_2
 
-          collection << mock_matches_1
-          collection.actual_potential_matches.should == mock_matches_array_1
+          subject << mock_matches_1
+          subject.actual_potential_matches.should == mock_matches_array_1
 
-          collection << mock_matches_2
-          collection.actual_potential_matches.should == mock_matches_array_1 + mock_matches_array_2
+          subject << mock_matches_2
+          subject.actual_potential_matches.should == mock_matches_array_1 + mock_matches_array_2
         end
 
         it "the <<() method should collect the exact matches into the actual_exact_matches Array" do
@@ -31,23 +36,22 @@ module Adhearsion
           mock_matches_1 = mock_with_exact_matches mock_matches_array_1
           mock_matches_2 = mock_with_exact_matches mock_matches_array_2
 
-          collection << mock_matches_1
-          collection.actual_exact_matches.should == mock_matches_array_1
+          subject << mock_matches_1
+          subject.actual_exact_matches.should == mock_matches_array_1
 
-          collection << mock_matches_2
-          collection.actual_exact_matches.should == mock_matches_array_1 + mock_matches_array_2
+          subject << mock_matches_2
+          subject.actual_exact_matches.should == mock_matches_array_1 + mock_matches_array_2
         end
 
         it "if any exact matches exist, the exact_match?() method should return true" do
-          collection << mock_with_exact_matches([1,2,3])
-          collection.exact_match?.should be true
+          subject << mock_with_exact_matches([1,2,3])
+          subject.exact_match?.should be true
         end
 
         it "if any potential matches exist, the potential_match?() method should return true" do
-          collection << mock_with_potential_matches([1,2,3])
-          collection.potential_match?.should be true
+          subject << mock_with_potential_matches([1,2,3])
+          subject.potential_match?.should be true
         end
-
       end
     end
   end